#1543: split userspace-dp screen + SYN-cookie runtime into siblings
Decompose userspace-dp/src/screen.rs (1420 LOC) into focused
sibling submodules under userspace-dp/src/screen/ per the
module/foo.rs convention. Pure code motion; no behavior change, no
allocation change, no public API change.

Layout:
  screen/mod.rs          - ScreenState orchestrator + pub(crate) re-exports
  screen/packet.rs       - ScreenPacketInfo, ScreenProfile, ScreenVerdict,
                           PROTO_* / TCP_* constants
  screen/syncookie.rs    - SynCookieCodec, SynCookieValidatedCache,
                           SipHash24 MAC primitive, all cookie types
  screen/rate.rs         - RateCounter (per-zone 1-second window)
  screen/stateless.rs    - Pure free-fn helpers: LAND, TCP flag screens,
                           ping-of-death, teardrop, ICMP fragment,
                           source-route. All #[inline].
  screen/scan.rs         - PortScanTracker, IpSweepTracker
  screen/session_limit.rs- SessionLimitTracker (per-IP session counts)
  screen/extract.rs      - extract_screen_info (allocation-free parser)
  screen/tests.rs        - Relocated from src/screen_tests.rs

Order of checks and drop-reason strings preserved byte-for-byte.
Side-effect ordering preserved: rate counters increment before
threshold check returns Drop; SYN-cookie take_valid runs before
syn_counters.increment so validated cookies bypass the rate
counter. ScreenState method set unchanged; afxdp/ external callers
unchanged via pub(crate) use re-exports.

Plan review:
- AGY round-1: PLAN-NEEDS-MINOR (3 minor fix-ups, all incorporated
  into plan v2: pub(crate) re-exports of SipHash24 +
  SynCookieValidatedCache under #[cfg(test)] because tests
  construct them directly; line-range typo).
- AGY round-2: PLAN-READY.
- Codex rounds 1/2/3: sandbox infra failure
  (codex-linux-sandbox binary path persistently broken across
  3 rescue retries). Proceeded under the Codex-stuck 3-of-4
  exception per Wave-5 rules with Claude SMR (HPC networking +
  CPU arch + SW design patterns) as the third independent
  attestation.
- Claude SMR: PLAN-READY.

Validation:
- TMPDIR=/dev/shm CARGO_TARGET_DIR=/dev/shm/cargo cargo build clean
  (release + dev).
- TMPDIR=/dev/shm CARGO_TARGET_DIR=/dev/shm/cargo cargo test --release:
  1506/1507 pass. The single failure
  (snat_contract_documents_current_fail_closed_runtime) is a
  pre-existing master breakage unrelated to this PR (verified by
  running the same test on a clean origin/master worktree at
  /home/ps/git/bpfrx/.claude/worktrees/1528-dpdk-mechanical-removal
  with the same failure).
- screen::tests subset: 83/83 pass, 5x flake clean.
- Go suite: passes across 30 packages.

No per-PR smoke per Wave-5 rules (AWAITING-BATCH-MERGE marker).

Closes #1543
